+++ libs/filaflat/include/filaflat/MaterialChunk.h
@@ -20,6 +20,7 @@
 
 #include <filament/MaterialChunkType.h>
 
+#include <filaflat/ChunkContainer.h>
 #include <filaflat/Unflattener.h>
 
 #include <private/filament/Variant.h>
@@ -27,10 +28,6 @@
 #include <tsl/robin_map.h>
 
 namespace filaflat {
-
-class BlobDictionary;
-class ChunkContainer;
-class ShaderBuilder;
 
 class MaterialChunk {
 public:
@@ -41,7 +38,7 @@ public:
     bool readIndex(filamat::ChunkType materialTag);
 
     // call this as many times as needed
-    bool getShader(ShaderBuilder& shaderBuilder,
+    bool getShader(ShaderContent& shaderContent,
             BlobDictionary const& dictionary,
             uint8_t shaderModel, filament::Variant variant, uint8_t stage);
 
@@ -53,11 +50,11 @@ private:
     tsl::robin_map<uint32_t, uint32_t> mOffsets;
 
     bool getTextShader(Unflattener unflattener,
-            BlobDictionary const& dictionary, ShaderBuilder& shaderBuilder,
+            BlobDictionary const& dictionary, ShaderContent& shaderContent,
             uint8_t shaderModel, filament::Variant variant, uint8_t ps);
 
     bool getSpirvShader(
-            BlobDictionary const& dictionary, ShaderBuilder& shaderBuilder,
+            BlobDictionary const& dictionary, ShaderContent& shaderContent,
             uint8_t shaderModel, filament::Variant variant, uint8_t stage);
 };