+++ RELEASE_NOTES.md
@@ -5,6 +5,13 @@ A new header is inserted each time a *tag* is created.
 
 ## main branch
 
+## v1.25.0
+
+- Vulkan: smol-v blobs are now 8-byte aligned within the filamat archive. [⚠️ **Recompile Materials**]
+- backend: added support for EGL on linux (headless)
+- uberz tool: add --append and --template arguments.
+- matc tool: add --template argument.
+
 ## v1.24.0
 
 - ImGuiHelper: add support for Y flip.